(int)pthread_self()

+++ ex3_q1_given.c
@@ -122,7 +122,7 @@ void print_grade_F(int indx){
 //------------------------------------------------------------------------
 void random_sleep()
 {
-    int d = (rand() ^ pthread_self() ^ getpid()) % 40 + 40;
+    int d = (rand() ^ (int)pthread_self() ^ getpid()) % 40 + 40;
     const struct timespec ts = {0, d * 1000 * 1000 * 5};
     nanosleep(&ts, NULL);
 }
